OT -PROTOTYPE/requirements.txt
fastapi
uvicorn[standard]
pyshark
pydantic
python-multipart
pyyaml